Added @scope("session") to controllers
I know we've talked using HttpSession objects, but after thoroughly
reading on the subject, the HttpSession object it's simple and how you
would do it in older Spring-MVC and other less capable frameworks, but
it's messy and it exposes your clean controller to the Servlet API and
needs null checking after you have called getAttribute method.
Meanwhile, scoping the Controllers with the session value, makes the
controller very clean, even more so if you scope an object and then
@autowire it to the @scope("request") controller (probably gonna do this
in the future). The downside, though, is that a new controller is
created for each session. This could bloat the session and in particular
could cause problems on a large scale system which is hosted on multiple
servers (but it's not the case here)